Working for multiple intersections
Trace faces; store next_v to access the face to be traced. 

At intersections, use orient2d to track when to stop tracing the face past the intersection, by checking if we are to the right or left of the segment. Not necessary for attachments, b/c they are vertical and thus can be tested using x coordinate alone.