Fixes to Myers to align with paper
In 4D, remove  Below(e), which is next, not prev.

In hash, if the x value of the index returned equals the intersection x value, then use the previous index. This ensures that endpoint intersections are processed in the current work index and not the next index (at which point the associated segment would have been removed.

As a result, it is not necessary to check separately for whether two endpoints are equal when reporting intersections.