Do not test vertices for is_outside
Because the sweep now intersects the circle limitedRadius at the end (and MyCW2 and MyCW3 intersect limitedAngle, etc.), cannot remove vertices based on being outside. If a vertex is removed but the line crosses the invisible boundary, the sweep polygon will be incorrect and thus will not result in a correct intersection shape with the boundary polygon or circle.

Instead, remove edges that are outside the boundary.